+++ mlflow/models/__init__.py
@@ -30,7 +30,7 @@ The built-in flavors are:
 
 For details, see `MLflow Models <../models.html>`_.
 """
-from mlflow.models.dependencies_schema import set_retriever_schema
+from mlflow.models.dependencies_schemas import set_retriever_schema
 from mlflow.models.evaluation import (
     EvaluationArtifact,
     EvaluationMetric,
