+++ mlflow/tracing/utils.py
@@ -190,7 +190,7 @@ def maybe_get_request_id(is_evaluate=False) -> Optional[str]:
 def maybe_get_dependencies_schema() -> Optional[dict]:
     context = _try_get_prediction_context()
     if context:
-        return context.dependencies_schema
+        return context.dependencies_schemas
 
 
 def traces_to_df(traces: List[Trace]) -> "pandas.DataFrame":
